PRO-2590: Re-add element as a unordered list and list item
A change was requested to make this a div with a paragraph because there
was only one item in the list.  However, the content that goes in that
list contains two hyperlinks, and therefore for accessibility reasons and
indeed to pass the Pa11y tests the element must be a ul and li.