Use new gen_id method in summary template.
This means that the elements can be reliably referenced. This approach means that
a screen reader will read 'Change xyz' or 'Answer xyz' rather than just 'Change'
or 'Answer'. This may not be ideal in welsh, but I don't know what our normal
interaction between i18n and a11y expectations are.