Use aria-label to provide a11y context for summary.
This now uses either a statically generated string from the link text
(which is Change or Answer in english) plus the text of the question.
This mostly works, but is a little clunky at times. Unclear if this
works in welsh though. Handling for specific text is available, but
unused at this time.