DTSPB-4869 webchat rework (#2635)
* Add webchat file as provided by kerv.

Obviously this does nothing currently.

* add commas at end of lists in app.js

* very rough but working webchat.

notably the text used is probably incorrect (refers to scotland, maybe
wrong phone number?) and does not translate to welsh, but as a proof of
life this is sufficient.

* use feature toggle to switch to new implementation.

* make tests pass with new context settings.

having tests that fail when unrelated stuff is added seems like a bad idea.

* add ft_use_new_webchat to all requests.

* use the value for the switch.

* toggle between deployments based on language.

this seems like a bad way to do this generally, but this should work well enough
for our testing to get started.

* load webchat config from secrets.

should permit different configurations on a per-environment basis.

* add secrets config to load for preview.

* add comment highlighting hard coded values in csp.

* use correct names for config vars.

* typo in deployment const.

* DTSPB-4909 update css and text

* DTSPB-4910 update chart and custom-environment-variables

* Pick up text from existing strings.

Added a couple for no JS enabled and for when status is being checked.

* translate popup message.

* minimize content in webchat include.

* make closed consistent with avaya, fix typo.

---------

Co-authored-by: FeliTam <feli.tam@justice.gov.uk>