Cludge for short term testing issue.
We do not currently have handling to correctly mock out the SubmitData calls here.
It appears that other tests are unintentionally passing the expected result array
into the handlePost call and then chcking the actual against it, but this means the
expected value is mutated by the method under test and thus always matches.

Fixing this properly would likely require work to mock the SubmitData and return the
correct values, but in service of getting this issue resolved we are just making the
test explicitly expect the error.